Move SeamlessInput and SeamlessTextArea components.
