+++ src/components/QuestionnaireDesign/index.js
@@ -4,8 +4,8 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 import CanvasSection from "./CanvasSection";
 import Canvas from "./Canvas";
-import SeamlessInput from "./SeamlessInput";
-import SeamlessTextArea from "./SeamlessTextArea";
+import SeamlessInput from "../SeamlessInput/SeamlessInput";
+import SeamlessTextArea from "../SeamlessTextArea/SeamlessTextArea";
 import Field from "components/Forms/Field";
 import Form from "components/Forms/Form";
 import CustomPropTypes from "custom-prop-types";
