+++ src/components/SeamlessInput/SeamlessInput.test.js
